requirements.txt
btcpay-python==1.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
ecdsa==0.16.1
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
requests==2.25.1
six==1.15.0
stem==1.8.0
urllib3==1.26.3
Werkzeug==1.0.1
